moved `getremotespids` call to user grpc client
