+++ gRPC/src/spidServer/proto_files/spidHandler.proto
@@ -11,7 +11,6 @@ service SpidHandler {
     rpc AddRemoteSpid (AddRemoteSpidRequest) returns (AddRemoteSpidResponse) {}
     rpc UpdateRemoteSpid (UpdateRemoteSpidRequest) returns (UpdateRemoteSpidResponse) {}
     rpc RemoveRemoteSpid (RemoveRemoteSpidRequest) returns (RemoveRemoteSpidResponse) {}
-    rpc GetRemoteSpids (GetRemoteSpidsRequest) returns (GetRemoteSpidsResponse) {}
 }
 
 message LockInfo {
@@ -87,13 +86,4 @@ message RemoveRemoteSpidRequest {
 
 message RemoveRemoteSpidResponse {
     string message = 1;
-}
-
-message GetRemoteSpidsRequest {
-    GlobalPosition position = 1;
-}
-
-message GetRemoteSpidsResponse {
-    string message = 1;
-    string marshaledSpids = 2;
 }
